+++ src/planner/planners/MatchSolver.cpp
@@ -7,6 +7,7 @@
 #include "planner/planners/MatchSolver.h"
 
 #include "visitor/RewriteMatchLabelVisitor.h"
+#include "util/ExpressionUtils.h"
 
 namespace nebula {
 namespace graph {
@@ -99,6 +100,97 @@ Expression *MatchSolver::doRewrite(const MatchAstContext *mctx, const Expression
     } else {
         return rewrite(labelExpr);
     }
+}
+
+Expression *MatchSolver::makeIndexFilter(const std::string &label,
+                                         const MapExpression *map,
+                                         QueryContext *qctx) {
+    auto &items = map->items();
+    Expression *root = new RelationalExpression(
+        Expression::Kind::kRelEQ,
+        new TagPropertyExpression(new std::string(label), new std::string(*items[0].first)),
+        items[0].second->clone().release());
+    for (auto i = 1u; i < items.size(); i++) {
+        auto *left = root;
+        auto *right = new RelationalExpression(
+            Expression::Kind::kRelEQ,
+            new TagPropertyExpression(new std::string(label), new std::string(*items[i].first)),
+            items[i].second->clone().release());
+        root = new LogicalExpression(Expression::Kind::kLogicalAnd, left, right);
+    }
+    return qctx->objPool()->add(root);
+}
+
+Expression *MatchSolver::makeIndexFilter(const std::string &label,
+                                         const std::string &alias,
+                                         const Expression *filter,
+                                         QueryContext *qctx) {
+    static const std::unordered_set<Expression::Kind> kinds = {Expression::Kind::kRelEQ,
+                                                               Expression::Kind::kRelLT,
+                                                               Expression::Kind::kRelLE,
+                                                               Expression::Kind::kRelGT,
+                                                               Expression::Kind::kRelGE};
+
+    std::vector<const Expression *> ands;
+    auto kind = filter->kind();
+    if (kinds.count(kind) == 1) {
+        ands.emplace_back(filter);
+    } else if (kind == Expression::Kind::kLogicalAnd) {
+        ands = ExpressionUtils::pullAnds(filter);
+    } else {
+        return nullptr;
+    }
+
+    std::vector<Expression *> relationals;
+    for (auto *item : ands) {
+        if (kinds.count(item->kind()) != 1) {
+            continue;
+        }
+
+        auto *binary = static_cast<const BinaryExpression *>(item);
+        auto *left = binary->left();
+        auto *right = binary->right();
+        const LabelAttributeExpression *la = nullptr;
+        const ConstantExpression *constant = nullptr;
+        if (left->kind() == Expression::Kind::kLabelAttribute &&
+            right->kind() == Expression::Kind::kConstant) {
+            la = static_cast<const LabelAttributeExpression *>(left);
+            constant = static_cast<const ConstantExpression *>(right);
+        } else if (right->kind() == Expression::Kind::kLabelAttribute &&
+                   left->kind() == Expression::Kind::kConstant) {
+            la = static_cast<const LabelAttributeExpression *>(right);
+            constant = static_cast<const ConstantExpression *>(left);
+        } else {
+            continue;
+        }
+
+        if (*la->left()->name() != alias) {
+            continue;
+        }
+
+        auto *tpExpr = new TagPropertyExpression(new std::string(label),
+                                                 new std::string(*la->right()->name()));
+        auto *newConstant = constant->clone().release();
+        if (left->kind() == Expression::Kind::kLabelAttribute) {
+            auto *rel = new RelationalExpression(item->kind(), tpExpr, newConstant);
+            relationals.emplace_back(rel);
+        } else {
+            auto *rel = new RelationalExpression(item->kind(), newConstant, tpExpr);
+            relationals.emplace_back(rel);
+        }
+    }
+
+    if (relationals.empty()) {
+        return nullptr;
+    }
+
+    auto *root = relationals[0];
+    for (auto i = 1u; i < relationals.size(); i++) {
+        auto *left = root;
+        root = new LogicalExpression(Expression::Kind::kLogicalAnd, left, relationals[i]);
+    }
+
+    return qctx->objPool()->add(root);
 }
 
 }  // namespace graph
